splitting off user to its own files
